RxJava Vertx refactor (#2531)
* Refactor

* Refactor
