+++ vertx-and-rxjava/pom.xml
@@ -10,7 +10,6 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.baeldung</groupId>
     <artifactId>vertx-and-rxjava</artifactId>
 
     <build>
