+++ vertx-and-rxjava/src/test/java/com/baeldung/weather/CityAndDayLength.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.weather;
 
 import java.text.MessageFormat;
 
@@ -7,12 +7,13 @@ class CityAndDayLength {
     private final String city;
     private final double dayLengthInHours;
 
-    public CityAndDayLength(String city, long dayLengthInSeconds) {
+    CityAndDayLength(String city, long dayLengthInSeconds) {
         this.city = city;
         this.dayLengthInHours = dayLengthInSeconds / (60.0 * 60.0);
     }
 
-    @Override public String toString() {
+    @Override
+    public String toString() {
         return MessageFormat.format("In {0} there are {1,number,#0.0} hours of light.", city, dayLengthInHours);
     }
 }
