+++ vertx-and-rxjava/src/test/java/com/baeldung/weather/MetaWeatherClient.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.weather;
 
 import io.reactivex.Flowable;
 import io.vertx.core.http.RequestOptions;
@@ -8,38 +8,38 @@ import io.vertx.reactivex.core.http.HttpClientResponse;
 
 import static java.lang.String.format;
 
-public class MetaWeatherClient {
+class MetaWeatherClient {
 
     private static RequestOptions metawether = new RequestOptions()
-                .setHost("www.metaweather.com")
-                .setPort(443)
-                .setSsl(true);
+      .setHost("www.metaweather.com")
+      .setPort(443)
+      .setSsl(true);
 
     /**
      * @return A flowable backed by vertx that automatically sends an HTTP request at soon as the first subscription is received.
      */
-private static Flowable<HttpClientResponse> autoPerformingReq(HttpClient httpClient, String uri) {
-    HttpClientRequest req = httpClient.get(new RequestOptions(metawether).setURI(uri));
-    return req.toFlowable()
-            .doOnSubscribe(subscription -> req.end());
-}
+    private static Flowable<HttpClientResponse> autoPerformingReq(HttpClient httpClient, String uri) {
+        HttpClientRequest req = httpClient.get(new RequestOptions(metawether).setURI(uri));
+        return req.toFlowable()
+          .doOnSubscribe(subscription -> req.end());
+    }
 
-static Flowable<HttpClientResponse> searchByCityName(HttpClient httpClient, String cityName) {
-    HttpClientRequest req = httpClient.get(
-        new RequestOptions()
+    static Flowable<HttpClientResponse> searchByCityName(HttpClient httpClient, String cityName) {
+        HttpClientRequest req = httpClient.get(
+          new RequestOptions()
             .setHost("www.metaweather.com")
             .setPort(443)
             .setSsl(true)
             .setURI(format("/api/location/search/?query=%s", cityName)));
-    return req
-            .toFlowable()
-            .doOnSubscribe(subscription -> req.end());
-}
+        return req
+          .toFlowable()
+          .doOnSubscribe(subscription -> req.end());
+    }
 
     static Flowable<HttpClientResponse> getDataByPlaceId(HttpClient httpClient, long placeId) {
         return autoPerformingReq(
-                httpClient,
-                format("/api/location/%s/", placeId));
+          httpClient,
+          format("/api/location/%s/", placeId));
     }
 
 }
